requirements.txt
numpy==1.17.4
https://download.pytorch.org/whl/cpu/torch-1.0.1.post2-cp37-cp37m-linux_x86_64.whl
Flask==2.0.1
Pillow==8.2.0
gunicorn == 20.1.0
